Add services
We're starting with the following 13 microservices — each of which is a
`webapi` project, unless mentioned otherwise in backticks in parentheses:

- admin: Administrative dashboard (`blazorserver`)
- auth: User authentication and authorization
- client: The main client app that users will view (`blazorwasm`)
- components: Blazor components for admin and client apps (`razorclasslib`)
- content: Management of content that a user will go on to interact with
- forums: Management of forums (memberships and privileges)
- gateway: API gateway for the client app
- lib: Class library for all app entities (`classlib`)
- messaging: API for messaging among users
- metrics: Vital statistics of the app
- notif: Notification service for users
- search: Search for users, forums, and/or content
- userman: User CRUD, such as sign-up and profile management

Each of the service projects are accompanied by xUnit test projects. All
the projects are initialized only, with all the boilerplate code
generated by .NET Core CLI. Only a reference of the source project is
added to the test project, for each source-test pair.